+++ exchangelib/services/__init__.py
@@ -36,6 +36,7 @@ from .get_searchable_mailboxes import GetSearchableMailboxes
 from .get_server_time_zones import GetServerTimeZones
 from .get_user_availability import GetUserAvailability
 from .get_user_oof_settings import GetUserOofSettings
+from .mark_as_junk import MarkAsJunk
 from .move_item import MoveItem
 from .resolve_names import ResolveNames
 from .send_item import SendItem
@@ -73,6 +74,7 @@ __all__ = [
     'GetServerTimeZones',
     'GetUserAvailability',
     'GetUserOofSettings',
+    'MarkAsJunk',
     'MoveItem',
     'ResolveNames',
     'SendItem',