+++ tests/test_items/test_messages.py
@@ -121,6 +121,23 @@ class MessagesTest(CommonItemTest):
         res = forward.save(self.account.drafts)
         self.account.bulk_delete([sent_item, reply, res])
 
+    def test_mark_as_junk(self):
+        # Test that we can mark a Message item as junk and non-junk, and that the message goes to the junk forlder and
+        # back to the the inbox.
+        item = self.get_test_item().save()
+        item.mark_as_junk(is_junk=False, move_item=False)
+        self.assertEqual(item.folder, self.test_folder)
+        self.assertEqual(self.test_folder.get(categories__contains=self.categories).id, item.id)
+        item.mark_as_junk(is_junk=True, move_item=False)
+        self.assertEqual(item.folder, self.test_folder)
+        self.assertEqual(self.test_folder.get(categories__contains=self.categories).id, item.id)
+        item.mark_as_junk(is_junk=True, move_item=True)
+        self.assertEqual(item.folder, self.account.junk)
+        self.assertEqual(self.account.junk.get(categories__contains=self.categories).id, item.id)
+        item.mark_as_junk(is_junk=False, move_item=True)
+        self.assertEqual(item.folder, self.account.inbox)
+        self.assertEqual(self.account.inbox.get(categories__contains=self.categories).id, item.id)
+
     def test_mime_content(self):
         # Tests the 'mime_content' field
         subject = get_random_string(16)